# backend/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ketches api server",
        "title": "Ketches API Server",
        "contact": {},
        "version": "v1"
    },
    "paths": {
        "/api/v1/apps": {
            "get": {
                "description": "List apps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "List Apps",
                "parameters": [
                    {
                        "type": "string",
                        "name": "envID",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageNo",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "projectID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ListAppsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "Create App",
                "parameters": [
                    {
                        "description": "App data",
                        "name": "app",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateAppRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AppModel"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/apps/refs": {
            "get": {
                "description": "Get all apps for refs under a specific environment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "All App Refs Under Environment",
                "parameters": [
                    {
                        "type": "string",
                        "name": "envID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.AppRef"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{appID}": {
            "get": {
                "description": "Get app by app ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "Get App",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AppModel"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "Update App",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated app data",
                        "name": "app",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateAppRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AppModel"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an app by app ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "Delete App",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{appID}/action": {
            "post": {
                "description": "Perform an action on an app (e.g., deploy, restart)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "App Action",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Action to perform on the app",
                        "name": "action",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AppActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AppModel"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{appID}/image": {
            "put": {
                "description": "Update the image of an app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "Update App Image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New app image",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateAppImageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AppModel"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{appID}/instances": {
            "get": {
                "description": "List instances of an app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "List App Instances",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.AppInstanceModel"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{appID}/instances/terminate": {
            "post": {
                "description": "Terminate a specific instance of an app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "Terminate App Instance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Terminate app instance request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TerminateAppInstanceRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{appID}/instances/{instanceName}/containers/{containerName}/exec": {
            "get": {
                "description": "Exec into a terminal of a specific container in an app instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "Exec App Container Terminal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Instance name",
                        "name": "instanceName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Container name",
                        "name": "containerName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request to execute command in container",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ExecAppContainerTerminalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/apps/{appID}/ref": {
            "get": {
                "description": "Get app ref by app ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "Get App Ref",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AppRef"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "model.AppRef": {
            "type": "object",
            "properties": {
                "appID": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "envID": {
                    "type": "string"
                },
                "projectID": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "model.CreateAppRequest": {
            "type": "object",
            "required": [
                "containerImage",
                "displayName",
                "envID",
                "replicas",
                "slug",
                "workloadType"
            ],
            "properties": {
                "containerImage": {
                    "type": "string"
                },
                "deploy": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "envID": {
                    "type": "string"
                },
                "limitCPU": {
                    "description": "in milliCPU (e.g., 1000 for 1 CPU, 2000 for 2 CPUs)",
                    "type": "integer"
                },
                "limitMemory": {
                    "description": "in MiB",
                    "type": "integer"
                },
                "registryPassword": {
                    "type": "string"
                },
                "registryUsername": {
                    "type": "string"
                },
                "replicas": {
                    "description": "Number of replicas for the app",
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "requestCPU": {
                    "description": "in milliCPU (e.g., 500 for 0.5 CPU, 1000 for 1 CPU)",
                    "type": "integer"
                },
                "requestMemory": {
                    "description": "in MiB",
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "workloadType": {
                    "type": "string",
                    "enum": [
                        "Deployment",
                        "StatefulSet"
                    ]
                }
            }
        },
...
        "model.ListAppsResponse": {
            "type": "object",
            "properties": {
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AppModel"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.UpdateAppImageRequest": {
            "type": "object",
            "required": [
                "containerImage"
            ],
            "properties": {
                "containerImage": {
                    "type": "string"
                },
                "registryPassword": {
                    "type": "string"
                },
                "registryUsername": {
                    "type": "string"
                }
            }
        },
        "model.UpdateAppRequest": {
            "type": "object",
            "required": [
                "displayName"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                }
            }
        },
...
    }
}